refactor(game-client): remove outdated CLEAN-UP-EXECUTION-PLAN.md and update routing for not-found and home pages
- Deleted the CLEAN-UP-EXECUTION-PLAN.md file to streamline documentation and remove obsolete content.
- Updated routing for the NotFound and Home pages to utilize new experience-based components, enhancing code organization and clarity.
- Introduced new Settings and Study pages with improved structure and localization support, ensuring better user experience and maintainability.
