+++ apps/game-client/src/app/[locale]/(translit)/translit/page.tsx
@@ -5,7 +5,6 @@ import {
   getMessagesForLocale,
   type RouteParamsWithLocalePromise,
 } from '@game-client/i18n';
-import { TranslitPageServer } from '@game-client/ui/screens/translit/translit-page-server';
 
 export function generateStaticParams() {
   return generateStaticParamsForSupportedLocales();
@@ -20,12 +19,4 @@ export async function generateMetadata({ params }: { params: RouteParamsWithLoca
   });
 }
 
-export default async function TranslitPage({
-  params,
-}: Readonly<{
-  params: RouteParamsWithLocalePromise;
-}>) {
-  const { locale } = await getLocalizedRouteParams(params);
-
-  return <TranslitPageServer locale={locale} />;
-}
+export { TranslitPage as default } from '@game-client/ui/experiences/translit/pages/translit-page';
