+++ apps/game-client/src/ui/components/panel/panel.stories.tsx
@@ -214,9 +214,6 @@ export const InvertedHeader: Story = {
   args: {
     variant: 'inverted',
   },
-  parameters: {
-    a11y: { test: 'todo' },
-  },
 };
 
 export const LeadingOnly: Story = {
