+++ apps/game-client/src/ui/experiences/not-found/components/not-found-client.tsx
