+++ apps/game-client/src/ui/experiences/not-found/pages/not-found-page.tsx
@@ -1,0 +1,5 @@
+import { NotFoundClient } from '../components/not-found-client';
+
+export function NotFoundPage() {
+  return <NotFoundClient />;
+}
