+++ apps/game-client/src/ui/experiences/settings/pages/settings-page.tsx
@@ -1,27 +1,35 @@
-import { getMessagesForLocale, type SupportedLocale } from '@game-client/i18n';
-import { getLibraryServer } from '@game-client/learning-content';
+import {
+  getLocalizedRouteParams,
+  getMessagesForLocale,
+  type RouteParamsWithLocalePromise,
+} from '@game-client/i18n';
 import { AppShell } from '@game-client/ui/components/layout/app-shell';
 import { GameClientAppBar } from '@game-client/ui/components/layout/game-client-app-bar';
 import { GameClientDock } from '@game-client/ui/components/layout/game-client-dock';
 import { RailPatternAlphabet } from '@game-client/ui/components/layout/rail-pattern-alphabet';
-import { TranslitClient } from './translit-client';
+import { SettingsClient } from '../components/settings-client';
 
-export async function TranslitPageServer({ locale }: Readonly<{ locale: SupportedLocale }>) {
+export async function SettingsPage({
+  params,
+}: Readonly<{
+  params: RouteParamsWithLocalePromise;
+}>) {
+  const { locale } = await getLocalizedRouteParams(params);
   const commonMessages = getMessagesForLocale(locale, 'common');
-  const library = await getLibraryServer(locale);
+
   return (
     <AppShell
       appBar={
         <GameClientAppBar
-          title={commonMessages.dock.translit}
+          title={commonMessages.dock.settings}
           eyeBrow={commonMessages.appBar.brand}
           backHref={`/${locale}/explore/alphabet`}
         />
       }
-      startRailContent={<GameClientDock activeItemId="translit" />}
+      startRailContent={<GameClientDock activeItemId="settings" />}
       endRailContent={<RailPatternAlphabet />}
     >
-      <TranslitClient library={library} />
+      <SettingsClient />
     </AppShell>
   );
 }
