+++ apps/game-client/src/ui/experiences/study/components/study-screen-swipe.test.ts
