+++ apps/game-client/src/ui/experiences/study/components/study-screen.test.tsx
