+++ apps/game-client/src/ui/experiences/study/components/study-screen.tsx
