+++ apps/game-client/src/ui/experiences/translit/components/translit-input.stories.tsx
@@ -2,11 +2,11 @@ import { Panel } from '@game-client/ui/components/panel/panel';
 import { PanelHeader } from '@game-client/ui/components/panel/panel-header';
 import { PanelSection } from '@game-client/ui/components/panel/panel-section';
 import { Surface } from '@game-client/ui/components/surface/surface';
-import { TranslitInput } from '@game-client/ui/screens/translit/translit-input';
 import { cn } from '@kartuli/ui/utils/cn';
 import type { Meta, StoryObj } from '@storybook/react-vite';
 import type { ComponentProps } from 'react';
 import { useEffect, useState } from 'react';
+import { TranslitInput } from './translit-input';
 
 function TranslitInputStoryRender(args: ComponentProps<typeof TranslitInput>) {
   const [value, setValue] = useState(args.value ?? '');
