+++ apps/game-client/src/ui/experiences/translit/components/translit-output-segments.ts
