+++ apps/game-client/src/ui/experiences/translit/components/translit-output.tsx
@@ -2,9 +2,9 @@
 
 import { Tooltip } from '@base-ui/react/tooltip';
 import { tooltipPopupClassName } from '@game-client/ui/components/tooltip';
-import type { TranslitOutputSegment } from '@game-client/ui/screens/translit/translit-output-segments';
 import { cn } from '@kartuli/ui/utils/cn';
 import { type RefObject, type UIEventHandler, useEffect, useRef, useState } from 'react';
+import type { TranslitOutputSegment } from './translit-output-segments';
 
 type TranslitOutputProps = {
   ariaLabelledBy: string;
