--- apps/game-client/src/ui/screens/home/home-page-server.tsx
@@ -1,15 +1,0 @@
-export function HomePageServer() {
-  return (
-    <main>
-      <h1>Kartuli is in cleanup mode</h1>
-      <p>
-        The old home screen has been removed so the new route structure can be rebuilt with less
-        noise.
-      </p>
-      <p>
-        Translit, settings, locale behavior, and the learning-content system are still in the repo.
-        The home route is intentionally a temporary placeholder for Phase 1.
-      </p>
-    </main>
-  );
-}
