--- apps/game-client/src/ui/screens/not-found/not-found-page-server.tsx
@@ -1,5 +1,0 @@
-import { NotFoundClient } from './not-found-client';
-
-export function NotFoundPageServer() {
-  return <NotFoundClient />;
-}
